+++ src/ru/yandex/practicum/delivery/model/parcels/PerishableParcel.java
@@ -5,7 +5,7 @@ import java.util.Date;
 public class PerishableParcel extends Parcel{
     int timeToLive;
 
-    public PerishableParcel(String description, double weight, String deliveryAddress, LocalDate sendDay, int timeToLive) {
+    public PerishableParcel(String description, int weight, String deliveryAddress, LocalDate sendDay, int timeToLive) {
         super(description, weight, deliveryAddress, sendDay, (double) 3);
         this.timeToLive = timeToLive;
     }
