+++ src/ru/yandex/practicum/delivery/model/parcels/StandardParcel.java
@@ -3,7 +3,7 @@ package ru.yandex.practicum.delivery.model.parcels;
 import java.time.LocalDate;
 
 public class StandardParcel extends Parcel {
-    public StandardParcel(String description, double weight, String deliveryAddress, LocalDate sendDay) {
+    public StandardParcel(String description, int weight, String deliveryAddress, LocalDate sendDay) {
         super(description, weight, deliveryAddress, sendDay, (double) 2);
     }
